+++ examples/tutorials/hybrid_swap_tutorial.py
@@ -100,9 +100,11 @@ class HybridSwapMCState(SwapMCState, MDState):
     from MDState.
     """
 
-    last_permutation: torch.Tensor
     _atom_attributes = (
-        MDState._atom_attributes | {"last_permutation"}  # noqa: SLF001
+        ts.SwapMCState._atom_attributes | MDState._atom_attributes  # noqa: SLF001
+    )
+    _system_attributes = (
+        ts.SwapMCState._system_attributes | MDState._system_attributes  # noqa: SLF001
     )