+++ torch_sim/__init__.py
@@ -8,6 +8,7 @@ from importlib.metadata import version
 import torch_sim as ts
 from torch_sim import (
     autobatching,
+    constraints,
     elastic,
     io,
     math,